backend/requirements.txt
google-generativeai>=0.8.0
google-auth>=2.35.0
google-cloud-firestore>=2.19.0
google-cloud-translate>=3.16.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
slowapi>=0.1.9
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
